Added sendIR, added byte macros, modified code to comply with standards

+++ src/biscuit.c
@@ -1,9 +1,9 @@
-bisc_ledOn(int led, int color, int intensity) {
+bisc_ledOn(uint8_t led, uint8_t color, uint8_t intensity) {
     bisc_core_led(led, color, intensity);
 }
 
 
-bisc_ledOff(int led) {
+bisc_ledOff(uint8_t led) {
     // Power LED = 0x08
     // Advance LED = 0x02
     // Both power and advance = 0x0A
