+++ src/biscuit.h
@@ -1,2 +1,2 @@
-bisc_ledOn(int led, int color, int intensity);
-bisc_ledOff(int led);
+bisc_ledOn(uint8_t led, uint8_t color, uint8_t intensity);
+bisc_ledOff(uint8_t led);
